Update to v0.1.8+4fix2

+++ lib/utils/frpc/process_manager.dart
@@ -51,10 +51,11 @@ class FrpcProcessManager {
     process.stdout.forEach((List<int> element) {
       final RegExp regex = RegExp(r'\x1B\[[0-9;]*[mK]');
       final String fmtStr = utf8.decode(element).trim().replaceAll(regex, '');
-      if (fmtStr.contains('[W]') ||
+      if ((fmtStr.contains('[W]') ||
           fmtStr.contains('stopped') ||
           fmtStr.contains('启动失败') ||
-          fmtStr.contains('无法连接')) {
+          fmtStr.contains('无法连接')) &&
+          !fmtStr.contains('重连失败')) {
         Logger.frpcWarn('[$proxyId] $fmtStr');
         fctr.appendWarnLog(fmtStr);
         process.kill();
